@@ -1,18 +1,20 @@
 {
   "name": "just-install",
-  "version": "1.0.11",
-  "description": "A Node.js installer for the excellent Just command runner (https://just.systems)",
+  "version": "2.0.0",
+  "description": "Install the Just command runner (https://just.systems) as part of `npm install` (Mac/Linux/Windows).",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "postinstall": "node ./install.js",
     "lint": "gts lint",
     "fix": "gts fix"
   },
   "bin": {
-    "just": "./bin/just"
+    "just": "./bin/just.js",
+    "just-install": "./install.js"
   },
   "engines": {
-    "node": ">=8.5"
+    "node": ">=18.0"
   },
   "repository": "https://github.com/brombal/just-install",
   "author": "Alex Brombal",
@@ -27,8 +29,10 @@
     "runner"
   ],
   "devDependencies": {
-    "gts": "^3.1.1",
-    "typescript": "~4.7.0",
-    "@types/node": "^14.11.2"
+    "gts": "^3.1.1"
+  },
+  "dependencies": {
+    "extract-zip": "^2.0.1",
+    "node-fetch": "^3.3.2"
   }
 }
