@@ -1,3 +1,4 @@
 module.exports = {
-  ...require('gts/.prettierrc.json')
-}
+  ...require('gts/.prettierrc.json'),
+  bracketSpacing: true,
+};
